Frontend/requirements.txt
streamlit==1.39.0
streamlit-pdf-viewer==1.0.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
firebase-admin==6.5.0
google-auth==2.27.0
